automatically detect and upgrade from local state v0 to v1 (#306)
* add ability to extend flights and formations

* load and upgrade state

This commit makes loading state attempt to start from the earliest
state, and upgrade up through to the current state.

The general idea is we optmistically load the current state, then
attempt to load previous states adding and deduping along the way.
Future state changes may require more sophisticated mechanisms for
transitioning.

This comes at the cost of some performance and complexity, so it may be
reduced in scope later if this becomes an issue (e.g. to current state
minus one, etc.).